requirements.txt
Flask==3.0.2
joblib==1.3.2
numpy==1.25.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
pandas==2.0.3
protobuf==4.25.3
scikit_learn==1.3.0
streamlit==1.33.0
websockets==12.0
yfinance==0.2.32
google.generativeai==0.4.1
streamlit_option_menu==0.3.12
